+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# main
+- Implement port enumeration for BSD and Apple platforms.
+- Implement port enumeration for Illumos and Solaris.
+
 # Version 0.1.0-alpha4 - 2021-11-09
 - Fix buffer truncation when enumerating ports on Windows.
 - Switch to overlapped IO on windows to allow concurrent reads and writes.